# docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: JARVIS Backend API
  version: 1.1.0
  description: |
    v1 API contract aligned with currently implemented Fastify routes.
servers:
  - url: https://api.jarvis.local
    description: Default server
...
paths:
...
  /api/v1/missions:
    post:
      tags: [Missions]
      summary: Create mission for multi-step orchestration
      operationId: createMission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionCreateRequest'
      responses:
        '201':
          description: Mission created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessMission'
        default:
          $ref: '#/components/responses/ApiError'
    get:
      tags: [Missions]
      summary: List missions
      operationId: listMissions
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, planned, running, blocked, completed, failed]
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Mission list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessMissions'
        default:
          $ref: '#/components/responses/ApiError'
  /api/v1/missions/{missionId}:
    get:
      tags: [Missions]
      summary: Get mission detail
      operationId: getMission
      parameters:
        - in: path
          name: missionId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mission detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessMission'
        '404':
          $ref: '#/components/responses/ApiError'
        default:
          $ref: '#/components/responses/ApiError'
    patch:
      tags: [Missions]
      summary: Update mission status and step statuses
      operationId: updateMission
      parameters:
        - in: path
          name: missionId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionUpdateRequest'
      responses:
        '200':
          description: Mission updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessMission'
        '404':
          $ref: '#/components/responses/ApiError'
        default:
          $ref: '#/components/responses/ApiError'
  /api/v1/missions/{missionId}/events:
    get:
      tags: [Events, Missions]
      summary: Stream mission updates over Server-Sent Events
      operationId: streamMissionEvents
      parameters:
        - in: path
          name: missionId
          required: true
          schema:
            type: string
        - in: query
          name: poll_ms
          schema:
            type: integer
            minimum: 300
            maximum: 10000
            default: 1200
        - in: query
          name: timeout_ms
          schema:
            type: integer
            minimum: 1000
            maximum: 120000
            default: 45000
      responses:
        '200':
          description: Mission event stream
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/MissionSseDataEnvelope'
        '404':
          $ref: '#/components/responses/ApiError'
        default:
          $ref: '#/components/responses/ApiError'
...
components:
  responses:
    ApiError:
      description: Standard API error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
  schemas:
    ApiMeta:
      type: object
      additionalProperties: true
    ApiErrorEnvelope:
      type: object
      required: [request_id, error]
      properties:
        request_id:
          type: string
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    MissionStep:
      type: object
      required: [id, type, title, description, route, status, order]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [code, research, finance, news, approval, execute]
        title:
          type: string
        description:
          type: string
        route:
          type: string
        status:
          type: string
          enum: [pending, running, done, blocked, failed]
        order:
          type: integer
          minimum: 1
    Mission:
      type: object
      required: [id, userId, workspaceId, title, objective, domain, status, steps, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        objective:
          type: string
        domain:
          type: string
          enum: [code, research, finance, news, mixed]
        status:
          type: string
          enum: [draft, planned, running, blocked, completed, failed]
        steps:
          type: array
          items:
            $ref: '#/components/schemas/MissionStep'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MissionCreateRequest:
      type: object
      required: [title, objective]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 180
        objective:
          type: string
          minLength: 1
          maxLength: 4000
        domain:
          type: string
          enum: [code, research, finance, news, mixed]
          default: mixed
        workspace_id:
          type: string
          format: uuid
        steps:
          type: array
          maxItems: 40
          items:
            type: object
            required: [type, title]
            properties:
              type:
                type: string
                enum: [code, research, finance, news, approval, execute]
              title:
                type: string
                minLength: 1
                maxLength: 180
              description:
                type: string
                maxLength: 2000
              route:
                type: string
                minLength: 1
                maxLength: 200
    MissionUpdateStepStatus:
      type: object
      required: [step_id, status]
      properties:
        step_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, done, blocked, failed]
    MissionUpdateRequest:
      type: object
      properties:
        status:
          type: string
          enum: [draft, planned, running, blocked, completed, failed]
        title:
          type: string
          minLength: 1
          maxLength: 180
        objective:
          type: string
          minLength: 1
          maxLength: 4000
        step_statuses:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/MissionUpdateStepStatus'
      anyOf:
        - required: [status]
        - required: [title]
        - required: [objective]
        - required: [step_statuses]
...
    MissionStreamOpen:
      type: object
      required: [request_id, mission_id]
      properties:
        request_id:
          type: string
        mission_id:
          type: string
    MissionStreamClose:
      type: object
      required: [request_id, mission_id]
      properties:
        request_id:
          type: string
        mission_id:
          type: string
        reason:
          type: string
    MissionStreamUpdate:
      type: object
      required: [mission_id, timestamp, data]
      properties:
        mission_id:
          type: string
        timestamp:
          type: string
          format: date-time
        data:
          $ref: '#/components/schemas/Mission'
    MissionSseDataEnvelope:
      oneOf:
        - $ref: '#/components/schemas/MissionStreamOpen'
        - $ref: '#/components/schemas/MissionStreamClose'
        - $ref: '#/components/schemas/MissionStreamUpdate'
...
    ApiSuccessMission:
      type: object
      required: [request_id, data, meta]
      properties:
        request_id:
          type: string
        data:
          $ref: '#/components/schemas/Mission'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ApiSuccessMissions:
      type: object
      required: [request_id, data, meta]
      properties:
        request_id:
          type: string
        data:
          type: object
          required: [missions]
          properties:
            missions:
              type: array
              items:
                $ref: '#/components/schemas/Mission'
        meta:
          $ref: '#/components/schemas/ApiMeta'
...